requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg[binary]==3.2.11
sqlalchemy==2.0.36
pydantic-settings==2.1.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
email-validator==2.1.1

# RAG System Dependencies
celery[redis]==5.3.4
redis==4.5.4
boto3==1.34.14
pinecone-client==5.0.1
# tiktoken==0.5.2  # Skipped for Phase 1 - requires Rust compiler on Python 3.14
python-magic==0.4.27
supabase==2.3.0
azure-storage-blob==12.19.0

# File Conversion Dependencies
openpyxl==3.1.2  # Excel file handling and preprocessing

# Document Parsing Dependencies
azure-ai-documentintelligence>=1.0.0b4  # Azure Document Intelligence (upgraded from Form Recognizer)
openai==1.54.4  # OpenAI for optional LLM enrichment
litellm>=1.52.0  # AI gateway for logging, guardrails, and model management
PyMuPDF>=1.23.8  # PDF image cropping for vision extraction (fitz)
Pillow>=10.2.0  # Image processing for vision extraction

# HTTP client dependencies (updated for Python 3.14 compatibility)
httpcore>=1.0.9  # Required for Python 3.14 compatibility
httpx>=0.28.1  # Required for Python 3.14 compatibility (may conflict with supabase, but needed for openai)

